+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 ## Unreleased
 
     * [SER-1406] - Fix create client errors
+    * [SER-842] - Allow an authorised user to select an active country so that s/he can make any changes to a specific country
 
 ## Version 1.0.0 - for use with Fineract Web App